+++ Translate5/MaintenanceCli/Command/T5Memory/T5MemoryReorganizeCommand.php
@@ -27,23 +27,28 @@
  */
 declare(strict_types=1);
 
-namespace Translate5\MaintenanceCli\Command;
+namespace Translate5\MaintenanceCli\Command\T5Memory;
 
 use editor_Models_LanguageResources_LanguageResource;
 use editor_Services_OpenTM2_Connector as Connector;
-use MittagQI\Translate5\Service\T5Memory\Enum\ReorganizeTm;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
+use Translate5\MaintenanceCli\Command\T5Memory\Traits\FilteringByNameTrait;
+use Translate5\MaintenanceCli\Command\T5Memory\Traits\T5MemoryLocalTmsTrait;
+use Translate5\MaintenanceCli\Command\Translate5AbstractCommand;
 use ZfExtended_Factory;
 
-final class T5MemoryReorganizeCommand extends T5memoryTmListCommand
+final class T5MemoryReorganizeCommand extends Translate5AbstractCommand
 {
+    use T5MemoryLocalTmsTrait;
+    use FilteringByNameTrait;
+
     protected static $defaultName = 't5memory:reorganize|memory:reorganize';
 
+    private const OPTION_TM_NAME = 'tmName';
     private const ARGUMENT_UUID = 'uuid';
 
     protected function configure(): void
@@ -55,7 +60,7 @@ final class T5MemoryReorganizeCommand extends T5memoryTmListCommand
             'UUID of the memory to reorganize, if not given, you can select from a list'
         );
         $this->addOption(
-            self::ARGUMENT_TM_NAME,
+            self::OPTION_TM_NAME,
             'f',
             InputOption::VALUE_REQUIRED,
             'If no UUID was given this will filter the list of all TMs if provided'
@@ -67,28 +72,13 @@ final class T5MemoryReorganizeCommand extends T5memoryTmListCommand
         $this->initInputOutput($input, $output);
         $this->initTranslate5();
 
+        $uuid = $this->getTmUuid($input);
 
-        $uuid = $input->getArgument(self::ARGUMENT_UUID);
-
-        if (empty($uuid)) {
-            $uuidList = $this->createLocalTmsUuidList();
-            if (empty($uuidList)) {
-                if ($this->input->hasOption(self::ARGUMENT_TM_NAME)) {
-                    $this->io->warning(
-                        'There are no translation memories that match "'
-                        . $this->input->getOption(self::ARGUMENT_TM_NAME)
-                        . '"');
-                } else {
-                    $this->io->warning('There are no translation memories in t5memory');
-                }
-                return self::FAILURE;
-            }
-            $askMemories = new ChoiceQuestion('Please choose a Memory:', array_values($uuidList), null);
-            $tmName = $this->io->askQuestion($askMemories);
-            $uuid = array_search($tmName, $uuidList);
+        if (null === $uuid) {
+            return self::FAILURE;
         }
 
-        $connector = $this->getConnector($uuid);
+        $connector = $this->getConnectorForLanguageResource($uuid);
 
         if (null === $connector) {
             return self::FAILURE;
@@ -152,7 +142,7 @@ final class T5MemoryReorganizeCommand extends T5memoryTmListCommand
         return $langResource;
     }
 
-    private function getConnector(string $languageResourceUuid): ?Connector
+    private function getConnectorForLanguageResource(string $languageResourceUuid): ?Connector
     {
         try {
             $languageResource = $this->getLanguageResource($languageResourceUuid);
@@ -178,12 +168,8 @@ final class T5MemoryReorganizeCommand extends T5memoryTmListCommand
         return $connector;
     }
 
-    private function createLocalTmsUuidList(): array
+    protected function getInput(): InputInterface
     {
-        $list = [];
-        foreach ($this->getLocalTms() as $item) {
-            $list[$item['uuid']] = $item['name'];
-        }
-        return $list;
+        return $this->input;
     }
 }